requirements.txt
pandas
pyarrow
requests
google-cloud-core
google-cloud-bigquery
python-dotenv